add wallet and transaction screens; implement WalletScreen, AllTransactionScreen, and related widgets for improved transaction management and display
