+++ lib/ui/screens/center/bottom_bar.dart
@@ -1,21 +1,11 @@
 import 'package:flutter/material.dart';
 import '../../../core/colors.dart';
-import '../home/all_transaction_screen.dart';
+import '../transaction/all/all_transaction_screen.dart';
 import '../home/home_screen.dart';
-import '../home/wallet_screen.dart';
+import '../wallet/wallet_screen.dart';
 
-class BottomBar extends StatefulWidget {
+class BottomBar extends StatelessWidget {
   const BottomBar({super.key});
-
-  @override
-  State<BottomBar> createState() => _BottomBarState();
-}
-
-class _BottomBarState extends State<BottomBar>
-    with SingleTickerProviderStateMixin {
-  late int _selectedIndex;
-  late AnimationController _controller;
-  late Animation<double> _animation;
 
   static const List<Widget> _widgetOptions = <Widget>[
     HomeScreen(),
@@ -24,58 +14,27 @@ class _BottomBarState extends State<BottomBar>
   ];
 
   @override
-  void initState() {
-    super.initState();
-    _selectedIndex = 0;
-    _controller = AnimationController(
-      duration: const Duration(milliseconds: 300),
-      vsync: this,
-    );
-    _animation = CurvedAnimation(
-      parent: _controller,
-      curve: Curves.easeInOut,
-    );
-    _controller.forward(from: 0.0); // Start the animation initially
-  }
-
-  void _onItemTapped(int index) {
-    setState(() {
-      _selectedIndex = index;
-      _controller.forward(from: 0.0);
-    });
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: FadeTransition(
-        opacity: _animation,
-        child: _widgetOptions.elementAt(_selectedIndex),
-      ),
-      bottomNavigationBar: Container(
-        color: Color(0xFFF4F4F4),
-        child: BottomNavigationBar(
-          type: BottomNavigationBarType.fixed,
-          backgroundColor: Colors.transparent,
-          elevation: 0,
-          items: const <BottomNavigationBarItem>[
-            BottomNavigationBarItem(icon: Icon(Icons.home), label: ''),
-            BottomNavigationBarItem(
-                icon: Icon(Icons.account_balance_wallet), label: ''),
-            BottomNavigationBarItem(icon: Icon(Icons.person), label: ''),
-          ],
-          currentIndex: _selectedIndex,
-          selectedItemColor: MyColors.primary,
-          unselectedItemColor: Colors.grey,
-          onTap: _onItemTapped,
-          showSelectedLabels: false,
-          showUnselectedLabels: false,
+    final screen = MediaQuery.of(context).size;
+    return DefaultTabController(
+      length: _widgetOptions.length,
+      child: Scaffold(
+        body: TabBarView(
+          children: _widgetOptions,
+        ),
+        bottomNavigationBar: Container(
+          height: screen.height * 0.085,
+          color: Color(0xFFF4F4F4),
+          child: TabBar(
+            tabs: const <Tab>[
+              Tab(icon: Icon(Icons.home)),
+              Tab(icon: Icon(Icons.account_balance_wallet)),
+              Tab(icon: Icon(Icons.person)),
+            ],
+            labelColor: MyColors.primary,
+            unselectedLabelColor: Colors.grey,
+            indicatorColor: Colors.transparent,
+          ),
         ),
       ),
     );
