+++ lib/ui/screens/home/widget/header_widget.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+import '../../../../core/colors.dart';
+
+class HeaderWidget extends StatelessWidget implements PreferredSizeWidget {
+  const HeaderWidget({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AppBar(
+      backgroundColor: MyColors.primary,
+      elevation: 0,
+      title: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text('Good afternoon,',
+              style: TextStyle(fontSize: 16, color: Colors.black54)),
+          Text('Enjelin Morgeana',
+              style: TextStyle(fontSize: 22, fontWeight: FontWeight.bold)),
+        ],
+      ),
+    );
+  }
+
+  @override
+  Size get preferredSize => Size.fromHeight(kToolbarHeight);
+}
