+++ lib/ui/screens/transaction/details/widget/text_row_widget.dart
@@ -1,0 +1,53 @@
+import 'package:flutter/material.dart';
+
+import '../../../../../core/colors.dart';
+
+class TextRowWidget extends StatelessWidget {
+  final IconData icon;
+  final String label;
+  final String value;
+
+  const TextRowWidget(
+      {super.key,
+      required this.icon,
+      required this.label,
+      required this.value});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 12.0),
+      child: Row(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Icon(icon, color: MyColors.primary, size: 22),
+          const SizedBox(width: 12),
+          Expanded(
+            child: Text(
+              label,
+              textAlign: TextAlign.left,
+              style: const TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.w600,
+                color: Colors.black87,
+              ),
+            ),
+          ),
+          Spacer(),
+          Expanded(
+            child: Text(
+              value,
+              textAlign: TextAlign.right,
+              overflow: TextOverflow.clip,
+              style: TextStyle(
+                fontSize: 16,
+                fontWeight: FontWeight.w500,
+                color: Colors.grey.shade700,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
