+++ lib/ui/screens/wallet/widget/wallet_header.dart
@@ -1,45 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import '../../../core/colors.dart';
-import '../../../providers/transaction_provider.dart';
-import '../../../utils/formater.dart';
-import '../../widgets/theme_container.dart';
-import 'add_transaction_screen.dart';
-import 'home_screen.dart';
 
-class WalletScreen extends StatelessWidget {
-  const WalletScreen({super.key});
+import '../../../../providers/transaction_provider.dart';
+import '../../../../utils/formater.dart';
+import '../../transaction/add&edit/add_transaction_screen.dart';
+
+class WalletHeader extends StatelessWidget {
+  const WalletHeader({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      extendBodyBehindAppBar: false,
-      appBar: AppBar(
-        backgroundColor: MyColors.primary,
-        title: const Text(
-          'Wallet',
-          style: TextStyle(
-            fontSize: 22,
-            fontWeight: FontWeight.bold,
-            color: Colors.white,
-          ),
-        ),
-        centerTitle: true,
-        // backgroundColor: Colors.teal,
-        elevation: 0,
-      ),
-      body: ThemeContainer(
-        child: Column(
-          children: [
-            _buildHeader(context),
-            const Expanded(child: TransactionHistory()),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildHeader(BuildContext context) {
     return Consumer<TransactionProvider>(
       builder: (context, provider, child) {
         return Container(
