+++ lib/ui/widgets/text_field.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/material.dart';
+
+class MyTextField extends StatelessWidget {
+  final TextEditingController controller;
+  final String label;
+  final IconData icon;
+  final bool isNumber;
+  final bool isDate;
+
+  const MyTextField({
+    super.key,
+    required this.controller,
+    required this.label,
+    required this.icon,
+    this.isNumber = false,
+    this.isDate = false,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(vertical: 8),
+      child: TextField(
+        controller: controller,
+        keyboardType: isNumber ? TextInputType.number : TextInputType.text,
+        readOnly: isDate,
+        onTap: isDate
+            ? () async {
+                DateTime? pickedDate = await showDatePicker(
+                  context: context,
+                  initialDate: DateTime.now(),
+                  firstDate: DateTime(2000),
+                  lastDate: DateTime(2100),
+                  builder: (context, child) {
+                    return Theme(
+                      data: Theme.of(context).copyWith(
+                        colorScheme: ColorScheme.light(
+                          primary: Colors.teal, // header background color
+                          onPrimary: Colors.white, // header text color
+                          onSurface: Colors.black, // body text color
+                        ),
+                        textButtonTheme: TextButtonThemeData(
+                          style: TextButton.styleFrom(
+                            foregroundColor: Colors.teal, // button text color
+                          ),
+                        ),
+                      ),
+                      child: child!,
+                    );
+                  },
+                );
+                if (pickedDate != null) {
+                  controller.text = "${pickedDate.toLocal()}".split(' ')[0];
+                }
+              }
+            : null,
+        decoration: InputDecoration(
+          labelText: label,
+          hintText: "Enter $label",
+          border: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(15),
+            borderSide: BorderSide(color: Colors.teal),
+          ),
+          enabledBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(15),
+            borderSide: BorderSide(color: Colors.teal),
+          ),
+          focusedBorder: OutlineInputBorder(
+            borderRadius: BorderRadius.circular(15),
+            borderSide: BorderSide(color: Colors.teal, width: 2),
+          ),
+          prefixIcon: Icon(icon, color: Colors.teal),
+          labelStyle: TextStyle(color: Colors.teal),
+          filled: true,
+          fillColor: Colors.white.withAlpha((0.1 * 255).toInt()),
+        ),
+        style: TextStyle(color: Colors.black),
+        cursorColor: Colors.teal,
+      ),
+    );
+  }
+}
